# Pfe/api-docs.json
{"openapi":"3.0.1","info":{"title":"STAR Amicale API","description":"PFE: Plateforme Gestion Amicale STAR Assurance","version":"Sprint 2"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"Bearer Authentication":[]}],"paths":{"/offres/modifier/{id}":{"put":{"tags":["offre-controller"],"operationId":"modifier","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateOffreRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OffreResponse"}}}}}}},"/inscriptions/{id}/paiement":{"put":{"tags":["inscription-controller"],"operationId":"mettreAjourPaiement","parameters":[],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InscriptionService"}}}}}}},"/admin/utilisateurs/{id}/desactiver":{"put":{"tags":["admin-controller"],"operationId":"desactiver","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/admin/utilisateurs/{id}/activer":{"put":{"tags":["admin-controller"],"operationId":"activer","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/offres/creer":{"post":{"tags":["offre-controller"],"operationId":"creer","parameters":[{"name":"titre","in":"query","required":true,"schema":{"type":"string"}},{"name":"description","in":"query","required":true,"schema":{"type":"string"}},{"name":"typeOffre","in":"query","required":true,"schema":{"type":"string","enum":["VOYAGE","EVENEMENT","CONVENTION"]}},{"name":"dateDebut","in":"query","required":true,"schema":{"type":"string"}},{"name":"dateFin","in":"query","required":true,"schema":{"type":"string"}},{"name":"capaciteMax","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"prixParPersonne","in":"query","required":true,"schema":{"type":"number","format":"double"}},{"name":"lieu","in":"query","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"image":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OffreResponse"}}}}}}},"/inscriptions/inscrire/{offreId}":{"post":{"tags":["inscription-controller"],"operationId":"inscrire","parameters":[{"name":"offreId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InscriptionResponse"}}}}}}},"/auth/refresh":{"post":{"tags":["auth-controller"],"operationId":"refresh","requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/auth/logout":{"post":{"tags":["auth-controller"],"operationId":"logout","responses":{"200":{"description":"OK"}}}},"/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/admin/utilisateurs":{"get":{"tags":["admin-controller"],"operationId":"listerTous","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"post":{"tags":["admin-controller"],"operationId":"creer_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUtilisateurRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/offres/fermer/{id}":{"patch":{"tags":["offre-controller"],"operationId":"fermer","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OffreResponse"}}}}}}},"/inscriptions/confirmer/{inscriptionId}":{"patch":{"tags":["inscription-controller"],"operationId":"confirmer","parameters":[{"name":"inscriptionId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InscriptionResponse"}}}}}}},"/inscriptions/annuler/{inscriptionId}":{"patch":{"tags":["inscription-controller"],"operationId":"annuler","parameters":[{"name":"inscriptionId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InscriptionResponse"}}}}}}},"/offres":{"get":{"tags":["offre-controller"],"operationId":"listerOuvertes","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OffreResponse"}}}}}}}},"/offres/{id}":{"get":{"tags":["offre-controller"],"operationId":"trouverParId","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OffreResponse"}}}}}}},"/offres/search":{"get":{"tags":["offre-controller"],"operationId":"rechercher","parameters":[{"name":"titre","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OffreResponse"}}}}}}}},"/inscriptions/offre/{offreId}":{"get":{"tags":["inscription-controller"],"operationId":"inscritsParOffre","parameters":[{"name":"offreId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/InscriptionResponse"}}}}}}}},"/inscriptions/mesinscriptions":{"get":{"tags":["inscription-controller"],"operationId":"mesInscriptions","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/InscriptionResponse"}}}}}}}},"/offres/supprimer/{id}":{"delete":{"tags":["offre-controller"],"operationId":"supprimer","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"UpdateOffreRequest":{"type":"object","properties":{"titre":{"type":"string"},"description":{"type":"string"},"typeOffre":{"type":"string","enum":["VOYAGE","EVENEMENT","CONVENTION"]},"dateDebut":{"type":"string","format":"date"},"dateFin":{"type":"string","format":"date"},"capaciteMax":{"type":"integer","format":"int32"},"prixParPersonne":{"type":"number","format":"double"},"lieu":{"type":"string"}}},"OffreResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"titre":{"type":"string"},"description":{"type":"string"},"typeOffre":{"type":"string","enum":["VOYAGE","EVENEMENT","CONVENTION"]},"statutOffre":{"type":"string","enum":["OUVERTE","FERMEE","ANNULEE"]},"dateDebut":{"type":"string","format":"date"},"dateFin":{"type":"string","format":"date"},"capaciteMax":{"type":"integer","format":"int32"},"placeRestantes":{"type":"integer","format":"int32"},"prixParPersonne":{"type":"number","format":"double"},"lieu":{"type":"string"},"imageBase64":{"type":"string"},"imageType":{"type":"string"},"imageNom":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"InscriptionService":{"type":"object"},"InscriptionResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"offreId":{"type":"integer","format":"int32"},"offreTitre":{"type":"string"},"adherentId":{"type":"integer","format":"int32"},"mailAdherent":{"type":"string"},"statut":{"type":"string","enum":["EN_ATTENTE","CONFIRMEE","ANNULEE"]},"statutPaiement":{"type":"string","enum":["EN_ATTENTE","VALIDE","REJETE"]},"montant":{"type":"number","format":"double"},"dateInscription":{"type":"string","format":"date-time"},"dateAnnulation":{"type":"string","format":"date-time"},"commentaire":{"type":"string"}}},"AuthResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"},"email":{"type":"string"},"role":{"type":"string"}}},"LoginRequest":{"required":["email","motDePasse"],"type":"object","properties":{"email":{"type":"string"},"motDePasse":{"type":"string"}}},"CreateUtilisateurRequest":{"required":["email","motDePasse","nom","prenom","role"],"type":"object","properties":{"email":{"type":"string"},"motDePasse":{"type":"string"},"nom":{"type":"string"},"prenom":{"type":"string"},"role":{"type":"string","enum":["ADMIN","ADHERENT","MEMBRE_BUREAU"]},"poste":{"type":"string"}}},"UserResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"email":{"type":"string"},"nom":{"type":"string"},"prenom":{"type":"string"},"role":{"type":"string"},"actif":{"type":"boolean"},"createdAt":{"type":"string","format":"date-time"}}}},"securitySchemes":{"Bearer Authentication":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}