# Pfe/src/main/java/tn/star/Pfe/dto/auth/UserResponse.java
package tn.star.Pfe.dto.auth;

import java.time.LocalDateTime;

public record UserResponse(
        int id,
        String email,
        String nom,
        String prenom,
        String role,
        boolean actif ,
        LocalDateTime createdAt) {
}